Move table initialization to page scripts.
