+++ templates/committees-single.html
@@ -205,5 +205,5 @@
 {% endblock %}
 
 {% block scripts %}
-  <script src="/static/js/vendor/tablist.js"></script>
+<script src="/{{ assets['dist/js/pages/committee-single.js'] }}"></script>
 {% endblock %}
