+++ templates/donations.html
@@ -38,3 +38,7 @@
 </div>
 
 {% endblock %}
+
+{% block scripts %}
+<script src="/{{ assets['dist/js/pages/receipts.js'] }}"></script>
+{% endblock %}
