+++ templates/expenditures.html
@@ -38,3 +38,7 @@
 </div>
 
 {% endblock %}
+
+{% block scripts %}
+<script src="/{{ assets['dist/js/pages/disbursements.js'] }}"></script>
+{% endblock %}
